Use more efficient Vec::push method for adding Strings to Vec
`Vec`s store their data in contiguous memory, which is generally
pre-allocated to some size and then extended where necessary to a larger
allocation. As such, it's generally relatively cheap to whack something
on the end of a `Vec`, using the `push` method, because most often you
just have to increment a pointer by the size of one item. If, however,
you want to insert an element anywhere else, that involves doing a copy
to move all the elements after the insertion point along by one, which
in general can be quite costly to do.

Another, more minor point, is that with your program the way you'd
written it, you'd get "buzz fizz" output for numbers that were divisible
by both 3 and 5, whereas the program is called "fizz_buzz", so it seems
more appropriate to have the keywords that way around! You could have
also fixed that by swapping the insertion order that you used, but that
would have still left the inefficiency of the inserts.