Completely change implementation for educational purposes
I'm _not_ saying that I like this code better, but I'm just showing you
another way to write the program so that you can weight the pros and
cons.

This version has the advantage that it removes all heap allocations, so
it should be super fast. Note that we bind `mod5` because we look at it
twice, so that we don't have duplicate code. In practice _the compiler
is smart enough to remove duplicate code like this for you_, so you
shouldn't need to worry about it (unlike in Python, which will just do
exactly what you say).

However, note that the logic is far less clear in this implementation,
and if someone comes along later and asks you to add another rule, this
is almost certainly a bad way to right it - you probably want the
previous version for maintainability!